+++ src/main/kotlin/com/github/mystery2099/woodenAccentsMod/block/custom/CoffeeTableBlock.kt
@@ -74,7 +74,7 @@ class CoffeeTableBlock(val baseBlock: Block, private val topBlock: Block) :
 
 
     override val tag: TagKey<Block> = ModBlockTags.coffeeTables
-    override val itemGroup = ModItemGroups.decorations
+    override val itemGroup = ModItemGroups.furniture
 
     private val BlockState.isTall: Boolean
         get() = getOrEmpty(type) == Optional.of(CoffeeTableTypes.TALL)
@@ -142,7 +142,7 @@ class CoffeeTableBlock(val baseBlock: Block, private val topBlock: Block) :
 
     @Deprecated("Deprecated in Java")
     override fun canReplace(state: BlockState, context: ItemPlacementContext): Boolean {
-        return !state.isTall && context.stack.item == asItem()
+        return !context.shouldCancelInteraction() && !state.isTall && context.stack.item == asItem()
     }
 
     private fun WorldAccess.checkDirection(pos: BlockPos, direction: Direction): Boolean {